Updated the plugin descriptor;
